Move sra_client module into zero_ex namespace
