+++ python-packages/sra_client/setup.py
@@ -97,20 +97,20 @@ class LintCommand(distutils.command.build_py.build_py):
 
     def run(self):
         """Run linter shell commands."""
+        lint_targets = "test src/zero_ex/sra_client/__init__.py setup.py"
         lint_commands = [
             # formatter:
             (
-                "black --line-length 79 --check --diff test"
-                " sra_client/__init__.py setup.py"
+                f"black --line-length 79 --check --diff test {lint_targets}"
             ).split(),
             # style guide checker (formerly pep8):
-            "pycodestyle test sra_client/__init__.py setup.py".split(),
+            f"pycodestyle {lint_targets}".split(),
             # docstring style checker:
-            "pydocstyle src test sra_client/__init__.py setup.py".split(),
+            f"pydocstyle {lint_targets}".split(),
             # static type checker:
-            "bandit -r test sra_client/__init__.py setup.py".split(),
+            f"bandit -r {lint_targets}".split(),
             # general linter:
-            "pylint test sra_client/__init__.py setup.py".split(),
+            f"pylint {lint_targets}".split(),
             # pylint takes relatively long to run, so it runs last, to enable
             # fast failures.
         ]
@@ -153,7 +153,9 @@ setup(
     url="https://github.com/0xproject/0x-monorepo/python-packages/sra_client",
     keywords=["OpenAPI", "OpenAPI-Generator", "Standard Relayer REST API"],
     install_requires=REQUIRES,
-    packages=find_packages(),
+    namespace_packages=["zero_ex"],
+    packages=find_packages("src"),
+    package_dir={"": "src"},
     include_package_data=True,
     long_description=README_MD,
     long_description_content_type="text/markdown",
@@ -186,7 +188,7 @@ setup(
     },
     command_options={
         "build_sphinx": {
-            "source_dir": ("setup.py", "."),
+            "source_dir": ("setup.py", "src"),
             "build_dir": ("setup.py", "build/docs"),
         }
     },
