--- python-packages/sra_client/sra_client/models/__init__.py
@@ -1,49 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-
-from __future__ import absolute_import
-
-# import models into model package
-from sra_client.models.order_schema import OrderSchema
-from sra_client.models.paginated_collection_schema import (
-    PaginatedCollectionSchema,
-)
-from sra_client.models.relayer_api_asset_data_pairs_response_schema import (
-    RelayerApiAssetDataPairsResponseSchema,
-)
-from sra_client.models.relayer_api_asset_data_trade_info_schema import (
-    RelayerApiAssetDataTradeInfoSchema,
-)
-from sra_client.models.relayer_api_error_response_schema import (
-    RelayerApiErrorResponseSchema,
-)
-from sra_client.models.relayer_api_error_response_schema_validation_errors import (
-    RelayerApiErrorResponseSchemaValidationErrors,
-)
-from sra_client.models.relayer_api_fee_recipients_response_schema import (
-    RelayerApiFeeRecipientsResponseSchema,
-)
-from sra_client.models.relayer_api_order_config_payload_schema import (
-    RelayerApiOrderConfigPayloadSchema,
-)
-from sra_client.models.relayer_api_order_config_response_schema import (
-    RelayerApiOrderConfigResponseSchema,
-)
-from sra_client.models.relayer_api_order_schema import RelayerApiOrderSchema
-from sra_client.models.relayer_api_orderbook_response_schema import (
-    RelayerApiOrderbookResponseSchema,
-)
-from sra_client.models.relayer_api_orders_channel_subscribe_payload_schema import (
-    RelayerApiOrdersChannelSubscribePayloadSchema,
-)
-from sra_client.models.relayer_api_orders_channel_subscribe_schema import (
-    RelayerApiOrdersChannelSubscribeSchema,
-)
-from sra_client.models.relayer_api_orders_channel_update_schema import (
-    RelayerApiOrdersChannelUpdateSchema,
-)
-from sra_client.models.relayer_api_orders_response_schema import (
-    RelayerApiOrdersResponseSchema,
-)
-from sra_client.models.signed_order_schema import SignedOrderSchema
