+++ python-packages/sra_client/src/zero_ex/__init__.py
@@ -1,0 +1,2 @@
+"""0x Python API."""
+__import__("pkg_resources").declare_namespace(__name__)
