+++ python-packages/sra_client/src/zero_ex/sra_client/__init__.py
@@ -52,7 +52,7 @@ the server.)
 Configure and create an API client instance
 -------------------------------------------
 
->>> from sra_client import ApiClient, Configuration, DefaultApi
+>>> from zero_ex.sra_client import ApiClient, Configuration, DefaultApi
 >>> config = Configuration()
 >>> config.host = "http://localhost:3000"
 >>> relayer_api = DefaultApi(ApiClient(config))
@@ -242,52 +242,50 @@ from __future__ import absolute_import
 __version__ = "1.0.0"
 
 # import apis into sdk package
-from sra_client.api.default_api import DefaultApi
+from .api.default_api import DefaultApi
 
 # import ApiClient
-from sra_client.api_client import ApiClient
-from sra_client.configuration import Configuration
+from .api_client import ApiClient
+from .configuration import Configuration
 
 # import models into sdk package
-from sra_client.models.order_schema import OrderSchema
-from sra_client.models.paginated_collection_schema import (
-    PaginatedCollectionSchema,
-)
-from sra_client.models.relayer_api_asset_data_pairs_response_schema import (
+from .models.order_schema import OrderSchema
+from .models.paginated_collection_schema import PaginatedCollectionSchema
+from .models.relayer_api_asset_data_pairs_response_schema import (
     RelayerApiAssetDataPairsResponseSchema,
 )
-from sra_client.models.relayer_api_asset_data_trade_info_schema import (
+from .models.relayer_api_asset_data_trade_info_schema import (
     RelayerApiAssetDataTradeInfoSchema,
 )
-from sra_client.models.relayer_api_error_response_schema import (
+from .models.relayer_api_error_response_schema import (
     RelayerApiErrorResponseSchema,
 )
-from sra_client.models.relayer_api_error_response_schema_validation_errors import (  # noqa: E501 (line too long)
+from .models.relayer_api_error_response_schema_validation_errors import (
     RelayerApiErrorResponseSchemaValidationErrors,
 )
-from sra_client.models.relayer_api_fee_recipients_response_schema import (
+from .models.relayer_api_fee_recipients_response_schema import (
     RelayerApiFeeRecipientsResponseSchema,
 )
-from sra_client.models.relayer_api_order_config_payload_schema import (
+from .models.relayer_api_order_config_payload_schema import (
     RelayerApiOrderConfigPayloadSchema,
 )
-from sra_client.models.relayer_api_order_config_response_schema import (
+from .models.relayer_api_order_config_response_schema import (
     RelayerApiOrderConfigResponseSchema,
 )
-from sra_client.models.relayer_api_order_schema import RelayerApiOrderSchema
-from sra_client.models.relayer_api_orderbook_response_schema import (
+from .models.relayer_api_order_schema import RelayerApiOrderSchema
+from .models.relayer_api_orderbook_response_schema import (
     RelayerApiOrderbookResponseSchema,
 )
-from sra_client.models.relayer_api_orders_channel_subscribe_payload_schema import (  # noqa: E501 (line too long)
+from .models.relayer_api_orders_channel_subscribe_payload_schema import (
     RelayerApiOrdersChannelSubscribePayloadSchema,
 )
-from sra_client.models.relayer_api_orders_channel_subscribe_schema import (
+from .models.relayer_api_orders_channel_subscribe_schema import (
     RelayerApiOrdersChannelSubscribeSchema,
 )
-from sra_client.models.relayer_api_orders_channel_update_schema import (
+from .models.relayer_api_orders_channel_update_schema import (
     RelayerApiOrdersChannelUpdateSchema,
 )
-from sra_client.models.relayer_api_orders_response_schema import (
+from .models.relayer_api_orders_response_schema import (
     RelayerApiOrdersResponseSchema,
 )
-from sra_client.models.signed_order_schema import SignedOrderSchema
+from .models.signed_order_schema import SignedOrderSchema
