+++ python-packages/sra_client/src/zero_ex/sra_client/api/__init__.py
@@ -3,4 +3,4 @@ from __future__ import absolute_import
 # flake8: noqa
 
 # import apis into api package
-from sra_client.api.default_api import DefaultApi
+from zero_ex.sra_client.api.default_api import DefaultApi
